+++ Library/include/actuators/Propeller.h
@@ -20,7 +20,7 @@
 //  Stonefish
 //
 //  Created by Patryk Cieslak on 06/05/2019.
-//  Copyright (c) 2019-2024 Patryk Cieslak. All rights reserved.
+//  Copyright (c) 2019-2025 Patryk Cieslak. All rights reserved.
 //
 
 #ifndef __Stonefish_Propeller__
@@ -45,10 +45,7 @@ namespace sf
          \param rightHand a flag to indicate if the propeller is right hand (clockwise rotation)
          \param inverted a flag to indicate if the setpoint is inverted (positive value results in backward force)
         */
-        Propeller(std::string uniqueName, SolidEntity* propeller, Scalar diameter, Scalar thrustCoeff, Scalar torqueCoeff, Scalar maxRPM, bool rightHand, bool inverted = false);
-        
-        //! A destructor.
-        ~Propeller();
+        Propeller(std::string uniqueName, std::shared_ptr<SolidEntity> propeller, Scalar diameter, Scalar thrustCoeff, Scalar torqueCoeff, Scalar maxRPM, bool rightHand, bool inverted = false);
         
         //! A method used to update the internal state of the thruster.
         /*!
@@ -94,7 +91,7 @@ namespace sf
         Scalar ki;
         Scalar iLim;
         Scalar omegaLim;
-        SolidEntity* prop;
+        std::shared_ptr<SolidEntity> propeller_;
         bool RH;
         bool inv;