+++ Tests/FluidDynamicsTest/FluidDynamicsTestManager.cpp
@@ -20,7 +20,7 @@
 //  Stonefish
 //
 //  Created by Patryk Cieslak on 01/07/2024.
-//  Copyright(c) 2024 Patryk Cieslak. All rights reserved.
+//  Copyright(c) 2024-2025 Patryk Cieslak. All rights reserved.
 //
 
 #include "FluidDynamicsTestManager.h"
@@ -32,6 +32,9 @@
 #include <entities/solids/Torus.h>
 #include <entities/solids/Cylinder.h>
 #include <entities/solids/Compound.h>
+#include <actuators/SimpleThruster.h>
+#include <sensors/scalar/Odometry.h>
+#include <sensors/Sample.h>
 #include <core/FeatherstoneRobot.h>
 #include <graphics/OpenGLTrackball.h>
 #include <utils/SystemUtil.hpp>
@@ -60,13 +63,18 @@ void FluidDynamicsTestManager::BuildScenario()
 
     ///////MATERIALS////////
     CreateMaterial("Light", sf::UnitSystem::Density(sf::CGS, sf::MKS, 0.9), 0.3);
+    CreateMaterial("Neutral", sf::UnitSystem::Density(sf::CGS, sf::MKS, 1.0), 0.5);
     CreateMaterial("Heavy", sf::UnitSystem::Density(sf::CGS, sf::MKS, 2.0), 0.9);
     SetMaterialsInteraction("Light", "Light", 0.5, 0.2);
     SetMaterialsInteraction("Heavy", "Heavy", 0.5, 0.2);
+    SetMaterialsInteraction("Neutral", "Neutral", 0.5, 0.5);
+    SetMaterialsInteraction("Light", "Neutral", 0.5, 0.2);
+    SetMaterialsInteraction("Heavy", "Neutral", 0.5, 0.2);
     SetMaterialsInteraction("Heavy", "Light", 0.5, 0.2);
     
     ///////LOOKS///////////
-    CreateLook("Yellow", sf::Color::RGB(1.f, 0.9f, 0.f), 0.3f, 0.f);
+    CreateLook("Yellow", sf::Color::RGB(1.f, 0.9f, 0.1f), 0.3f, 0.f);
+    CreateLook("Red", sf::Color::RGB(1.f, 0.1f, 0.1f), 0.3f, 0.f);
     
     ////////OBJECTS    
     EnableOcean(0.0);
@@ -85,7 +93,7 @@ void FluidDynamicsTestManager::BuildScenario()
     AddSolidEntity(sphere, sf::Transform(sf::IQ(), sf::Vector3(0, 0, 0)));
 
     //Box
-    sf::Vector3 boxDims(0.1, 0.2, 0.5);
+    sf::Vector3 boxDims(1.0, 0.5, 0.2);
     sf::Box* box = new sf::Box("Box", phy, boxDims, sf::I4(), "Light", "Yellow");
     AddSolidEntity(box, sf::Transform(sf::IQ(), sf::Vector3(2, 0, 0)));
 
@@ -111,7 +119,46 @@ void FluidDynamicsTestManager::BuildScenario()
     robot->DefineLinks(link1);
     robot->BuildKinematicStructure();
     AddRobot(robot, sf::Transform(sf::IQ(), sf::Vector3(4, 0, 0)));
-} 
+
+    // Drag testing
+    // Single solid entities     
+    sf::Box* hull1 = new sf::Box("Hull1", phy, boxDims, sf::I4(), "Neutral", "Yellow");
+    AddSolidEntity(hull1, sf::Transform(sf::IQ(), sf::Vector3(-7, 0, 1)));
+
+    sf::Box* hull2 = new sf::Box("Hull2", phy, boxDims, sf::I4(), "Neutral", "Yellow");
+    AddSolidEntity(hull2, sf::Transform(sf::Quaternion(M_PI_2, 0, 0), sf::Vector3(-5, 2, 1)));
+
+    std::shared_ptr<sf::Box> propeller = std::make_shared<sf::Box>("Propeller", phy, sf::Vector3(0.02,0.2,0.05), sf::I4(), "Neutral", "Red");
+    
+    sf::SimpleThruster* thruster1 = new sf::SimpleThruster("Thruster1", propeller, true, false);
+    thruster1->AttachToSolid(hull1, sf::Transform(sf::IQ(), sf::Vector3(-0.6, 0, 0)));
+    thruster1->setSetpoint(10.0, 0.0);
+    AddActuator(thruster1);
+
+    sf::SimpleThruster* thruster2 = new sf::SimpleThruster("Thruster2", propeller, true, false);
+    thruster2->AttachToSolid(hull2, sf::Transform(sf::IQ(), sf::Vector3(-0.6, 0, 0)));
+    thruster2->setSetpoint(10.0, 0.0);
+    AddActuator(thruster2);
+
+    // Compound solid entity
+    for (int i = 0; i < 8; ++i)
+    {
+        sf::Box* part1 = new sf::Box("Part1_" + std::to_string(i), phy, boxDims, sf::I4(), "Neutral", "Yellow");
+        sf::Box* part2 = new sf::Box("Part2_" + std::to_string(i), phy, boxDims, sf::I4(), "Neutral", "Yellow");
+        sf::Compound* comp = new sf::Compound("Comp" + std::to_string(i), phy, part1, sf::Transform(sf::IQ(), sf::Vector3(0, 0.5, 0)));
+        comp->AddExternalPart(part2, sf::Transform(sf::IQ(), sf::Vector3(0, -0.5, 0)));
+        AddSolidEntity(comp, sf::Transform(sf::Quaternion(M_PI_4*i, 0.0, 0.0), sf::Vector3(0, 0, 2 + 2*i)));
+
+        sf::Odometry* odometry = new sf::Odometry("Odometry" + std::to_string(i));
+        odometry->AttachToSolid(comp, sf::Transform(sf::IQ(), sf::Vector3(0, 0, 0)));
+        AddSensor(odometry);
+
+        sf::SimpleThruster* thruster = new sf::SimpleThruster("ThrusterComp" + std::to_string(i), propeller, true, false);
+        thruster->AttachToSolid(comp, sf::Transform(sf::IQ(), sf::Vector3(-0.6, 0, 0)));
+        thruster->setSetpoint(20.0, 0.0);
+        AddActuator(thruster);
+    }
+}
 
 void FluidDynamicsTestManager::SimulationStepCompleted(sf::Scalar timeStep)
 {
@@ -132,11 +179,11 @@ void FluidDynamicsTestManager::SimulationStepCompleted(sf::Scalar timeStep)
     sf::TestScalar("Sphere mass", sphere->getMass(), expectedSphereMass);
     sf::TestScalar("Sphere volume", sphere->getVolume(), expectedSphereVolume);
     sf::TestVector3("Sphere inertia", sphereInertia, sf::Vector3(expectedSphereInertia, expectedSphereInertia, expectedSphereInertia));
-    sf::TestScalar("Sphere submerged volume", sphere->getSubmergedVolume(), expectedSphereSubmergedVolume);
+    sf::TestScalar("Sphere submerged volume", sphere->getSubmergedVolume(), expectedSphereSubmergedVolume, 1e-3);
     
     //BOX
     sf::Box* box = (sf::Box*)getEntity("Box");
-    sf::Vector3 boxDims(0.1, 0.2, 0.5);
+    sf::Vector3 boxDims(1.0, 0.5, 0.2);
     sf::Vector3 boxInertia = box->getInertia();
 
     sf::Scalar expectedBoxVolume = boxDims.getX() * boxDims.getY() * boxDims.getZ();
@@ -148,7 +195,7 @@ void FluidDynamicsTestManager::SimulationStepCompleted(sf::Scalar timeStep)
     sf::TestScalar("Box mass", box->getMass(), expectedBoxMass);
     sf::TestScalar("Box volume", box->getVolume(), expectedBoxVolume);
     sf::TestVector3("Box inertia", boxInertia, expectedBoxInertia);
-    sf::TestScalar("Box submerged volume", box->getSubmergedVolume(), expectedBoxSubmergedVolume);
+    sf::TestScalar("Box submerged volume", box->getSubmergedVolume(), expectedBoxSubmergedVolume, 1e-3);
 
     //POLYHEDRON
     sf::Polyhedron* hull = (sf::Polyhedron*)getEntity("Hull");
@@ -163,8 +210,6 @@ void FluidDynamicsTestManager::SimulationStepCompleted(sf::Scalar timeStep)
     sf::TestScalar("Hull volume", hull->getVolume(), expectedHullVolume);
     sf::TestVector3("Hull inertia", hullInertia, expectedHullInertia);
     sf::TestScalar("Hull submerged volume", hull->getSubmergedVolume(), expectedHullSubmergedVolume);
-    printf("Hull CG transform:\n");
-    sf::PrintTransform(cgTransform, 5);
     
     //POLYHEDRON 2
     sf::Polyhedron* sphR1M = (sf::Polyhedron*)getEntity("SphereR1M");
@@ -176,10 +221,27 @@ void FluidDynamicsTestManager::SimulationStepCompleted(sf::Scalar timeStep)
     sf::Vector3 expectedSphR1MInertia = 2.0 / 5.0 * expectedSphR1MMass * sf::Vector3(1, 1, 1);
     sf::Scalar expectedSphR1MSubmergedVolume = expectedSphR1MVolume * sphR1M->getMaterial().density/rho;
 
-    sf::TestScalar("SphereR1M mass", sphR1M->getMass(), expectedSphR1MMass);
-    sf::TestScalar("SphereR1M volume", sphR1M->getVolume(), expectedSphR1MVolume);
-    sf::TestVector3("SphereR1M inertia", sphR1MInertia, expectedSphR1MInertia);
-    sf::TestScalar("SphereR1M submerged volume", sphR1M->getSubmergedVolume(), expectedSphR1MSubmergedVolume);
-    printf("SphereR1M CG transform:\n");
-    sf::PrintTransform(sphR1MCGTransform, 5);
+    sf::TestScalar("SphereR1M mass", sphR1M->getMass(), expectedSphR1MMass, 1e-2);
+    sf::TestScalar("SphereR1M volume", sphR1M->getVolume(), expectedSphR1MVolume, 1e-2);
+    sf::TestVector3("SphereR1M inertia", sphR1MInertia, expectedSphR1MInertia, 1e-2);
+    sf::TestScalar("SphereR1M submerged volume", sphR1M->getSubmergedVolume(), expectedSphR1MSubmergedVolume, 1e-2);
+
+    // DRAG TESTING
+    auto v1 = static_cast<sf::SolidEntity*>(getEntity("Hull1"))->getLinearVelocity();
+    auto v2 = static_cast<sf::SolidEntity*>(getEntity("Hull2"))->getLinearVelocity();
+
+    sf::TestScalar("Solid body drag", v1.length(), v2.length(), 0.01);
+
+    auto odometry = getSensor("Odometry0");
+    auto s = static_cast<sf::Odometry*>(odometry)->getLastSample();
+    sf::Scalar velX = s.getValue(3);
+
+    for (int i = 1; i < 8; ++i)
+    {
+        auto odometry = getSensor("Odometry" + std::to_string(i));
+        auto s = static_cast<sf::Odometry*>(odometry)->getLastSample();   
+        sf::TestScalar("Comp" + std::to_string(i) + " velocity", s.getValue(3), velX, 0.01);
+    }
+
+    std::cout << std::endl;
 }