+++ Tests/UnderwaterTest/UnderwaterTestManager.cpp
@@ -206,14 +206,14 @@ void UnderwaterTestManager::BuildScenario()
     //Create thrusters
     std::array<std::string, 5> thrusterNames = {"ThrusterSway", "ThrusterSurgePort", "ThrusterSurgeStarboard", "ThrusterHeaveStern", "ThrusterHeaveBow"}; 
     std::array<sf::Thruster*, 5> thrusters;
+    std::shared_ptr<sf::Polyhedron> propeller = std::make_shared<sf::Polyhedron>("Propeller", phy, sf::GetDataPath() + "propeller.obj", sf::Scalar(1), sf::I4(), "Dummy", "propeller");
     for(size_t i=0; i<thrusterNames.size(); ++i)
     {
-        sf::Polyhedron* prop = new sf::Polyhedron("Propeller", phy, sf::GetDataPath() + "propeller.obj", sf::Scalar(1), sf::I4(), "Dummy", "propeller");
         std::shared_ptr<sf::MechanicalPI> rotorDynamics;
         rotorDynamics = std::make_shared<sf::MechanicalPI>(1.0, 10.0, 5.0, 5.0);
         std::shared_ptr<sf::FDThrust> thrustModel;
         thrustModel = std::make_shared<sf::FDThrust>(0.18, 0.48, 0.48, 0.05, true, getOcean()->getLiquid().density);
-        thrusters[i] = new sf::Thruster(thrusterNames[i], prop, rotorDynamics, thrustModel, 0.18, true, 105.0, false, true);
+        thrusters[i] = new sf::Thruster(thrusterNames[i], propeller, rotorDynamics, thrustModel, 0.18, true, 105.0, false, true);
     }
 
     //Create VBS